Fix non-obvious index.html issues (#13)
When I was experimenting with different types of snapshotting/baselining behavior, I wanted it to be obvious to a non-expert user which detected failures were actually intentional. I thought it was a little more obvious to an unfamiliar reader for the 3 list items that are obviously examples to be the only examples caught by the tests.

This fixes the region and landmark-one-main violations, leaving exactly the one violation per bullet point that the bullet points self-describe.