Add comments in azure-pipelines.yml based on initial user feedback (#22)
One of the initial users of the sample stumbled a bit on the "Azure Pipelines integration" part. This change updates the comments in the sample to add more context for users that aren't very familiar with Azure Pipelines.

Also makes it a bit more clear that "npm" vs "yarn" doesn't matter, similar to what we did in the readme already.