chore: add Dependabot config file (#60)
This repo has multiple "root" folders with different languages, so it requires a checked in config file for Dependabot to work correctly.

For docs on this config file, see https://dependabot.com/docs/config-file

I think the only particularly interesting/subjective part of this config is the `version_requirement_updates: increase_versions` on the typescript folder. Essentially, we have to decide whether Dependabot should update package.json in the way it normally would for a library (where it would update yarn.lock but widen version ranges in package.json) or the way it normally would for an app (where it updates package.json to require the newer version). I think the latter makes more sense for a sample but could probably be persuaded otherwise.

I've already run this through the [Dependabot config validator](https://dependabot.com/docs/config-file/validator/)