[c#] Address user study feedback in SampleTests.cs (#105)
Addresses a variety of feedback items received in the recent round of user studies for the C# sample. Specifically:

* Moves the test setup/cleanup below the example test methods
* Clarifies which `using` statements a reader is likely to care about
* Adds a TestMethod demonstrating a basic scan of an entire page as the first example test, before the "scan one element in page" example
* Ensures that there is an example test that actually fails in practice, to give a better idea of what a failure would look like
* Clarify how `Include` works before using it incidentally alongside `WithTags`
* Add a more precise/verbose description of what "tags" and "rules" are and how to find more information about them
* clarify that Include/Exclude accept CSS selectors (not just IDs)
* Update sample page element IDs to make it more obvious on first reading that they are, in fact, IDs
* Switches order of see it in action locally/in pipelines sections of README

The most interesting of these changes is adding a test case that shows an example of what failure looks like. I wanted to achieve a balance where doing this wouldn't lead a real user cloning the sample project to include infrastructure for this that they wouldn't normally want. To achieve this, I:
* Had the new "basic scan of the whole page" test (which fails naturally) be allowed to fail
* Gave it a `[TestCategory("IntentionallyFailsAsAnExample")]`
* Set up `azure-pipelines.yml` to pass through a queue-time `EXTRA_DOTNET_TEST_ARGUMENTS` parameter to the dotnet test command
* Cloned the pipelines configuration so we now have 2 PR/CI builds that run the same yml file
* Named one `[failing example]` using no extra dotnet test arguments, and named one `[passing example]` using an extra dotnet test argument of `--filter TestCategory!=IntentionallyFailsAsAnExample`

This way, if a user cloned the sample without reading any of the comments and reused stuff blindly, the most unnecessary extra bits they'd get are an unnecessary benign `[TestCategory]` in one test method and an unnecessary benign `arguments: EXTRA_DOTNET_TEST_ARGUMENTS` property in the configuration of the dotnet test task in their pipelines yml.