chore(deps): bump playwright and @playwright/test to 1.13.1 in sync (#604)
#### Details

This PR updates `playwright` and `@playwright/test` in sync to 1.13.1. It also includes a bit of manual lockfile regeneration required to keep them in sync. This is required because the two packages include redundant sets of typings; our tests get passed types from the latter but @axe-core/playwright expects types from the former, so the build breaks if they get updated independently.

##### Motivation

Keep deps up to date

##### Context

Obsoletes #602  and #601 

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->

- [n/a] If this PR addresses an existing issue, it is linked: Fixes #0000
- [n/a] New sample content is commented at a similar verbosity as existing content
- [x] All updated/modified sample code builds and runs in at least one PR/CI build
- [ ] PR checks for builds named `[failing example] ...` fail as expected