chore: Allow dependabot to come up all green (#1237)
#### Details

PR's of this repo contain 2 checks that intentionally fail to
demonstrate what customers would do. These also fail in dependabot PR's,
so we need to take time to ensure not only that they failed, but that
they failed in the expected way. This only takes a minute or two for
each PR, but when we get 20 or more dependabot PR's per month, it starts
to add up. This PR leaves the failing behavior intact when running
locally or for non-dependabot PR's, but takes advantage of an ADO
pipeline variable to recognize a PR from dependabot and fail the check
only if an unexpected error occurs.

##### Motivation

Simplify dependabot PR handling

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [ ] If this PR addresses an existing issue, it is linked: Fixes #0000
- [ ] New sample content is commented at a similar verbosity as existing
content
- [ ] All updated/modified sample code builds and runs in at least one
PR/CI build
- [ ] PR checks for builds named `[failing example] ...` fail as
expected