chore: Make "failing" pipelines pass in PR builds but fail in CI builds (#1276)
#### Details

When first written, we had 2 pipelines that were "failing examples" to
show customers what failures look like. However, these pipelines created
a lot of friction for dependabot changes, so we added #1237 to
streamline the PR approval process. Unfortunately, the user-facing docs
use the pipeline as a demo of the failing builds, so this caused a
documentaion problem.

One option would be to split the PR and the CLI into 2 distinct
pipelines, but that introduces some organizational friction that we'd
prefer to avoid. The alternative is to use the same pipeline for both
the PR and CI builds, but make them behave differently--tests in PR
builds all pass (to streamline the review), while tests in CI builds
(i.e., built from the main branch) fail (to provide the doc examples).

The `ACCESSIBILITY_ERRORS_ARE_EXPECTED_IN_PR_BUILD` variable has already
been set in the 2 failing pipelines, so the unit tests for this PR
should come up all green.

##### Motivation

Make the docs useful while still streamlining the PR process

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [n/a] If this PR addresses an existing issue, it is linked: Fixes
#0000
- [n/a] New sample content is commented at a similar verbosity as
existing content
- [x] All updated/modified sample code builds and runs in at least one
PR/CI build
- [x] PR checks for builds named `[failing example] ...` pass PR's, then
confirm they fail in the CI build