chore: Disable combined PR's (#1319)
#### Details

Disable the combined dependabot PR's because we aren't using them.
If/when we ever decide to bring them back, it's trivial to revert the
change.

##### Motivation

Don't create extra branches that we aren't using.

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [n/a] If this PR addresses an existing issue, it is linked: 
- [ ] New sample content is commented at a similar verbosity as existing
content
- [ ] All updated/modified sample code builds and runs in at least one
PR/CI build
- [ ] PR checks for builds named `[failing example] ...` pass in the PR
build, then confirm they fail in the CI build